third_party/blink/web_tests/media/W3C/audio/preload/preload_reflects_none_autoplay-expected.txt

audio.preload - reflection test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "none" is "none"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
